pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rl-kg-agent"
version = "0.1.0"
description = "Reinforcement Learning Agent for Knowledge Graph Reasoning"
authors = [{name = "User", email = "user@example.com"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.20.0",
    "sentence-transformers>=2.2.0",
    "datasets>=2.0.0",
    "rdflib>=6.0.0",
    "SPARQLWrapper>=2.0.0",
    "stable-baselines3>=2.0.0",
    "gymnasium>=0.28.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "tqdm>=4.60.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.1",
    "torchrl>=0.6.0",
    "tensordict>=0.6.2",
    "playwright>=1.48.0",
    "matplotlib>=3.7.5",
    "tk>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-core>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
visualization = [
    "matplotlib>=3.7.0",
]

[project.scripts]
rl-kg-agent = "rl_kg_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rl_kg_agent"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src/rl_kg_agent",
    "/README.md",
    "/LICENSE",
    "/pyproject.toml",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py38"
line-length = 88
select = ["E", "F", "W", "B", "I"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
